+++ tests/test_request/test_bot.py
@@ -1,17 +1,14 @@
-import requests_mock
-
 import tgcli.request.bot
+
+import requests_mock
 
 
 class TestBotSession:
-    def setup_method(self):
-        self.session = tgcli.request.bot.BotSession("0")
+    def test_has_token(self, bot_session):
+        assert hasattr(bot_session, "token")
 
-    def test_has_token(self):
-        assert hasattr(self.session, "token")
-
-    def test_has_base_url(self):
-        assert hasattr(self.session, "base_url")
+    def test_has_base_url(self, bot_session):
+        assert hasattr(bot_session, "base_url")
 
 
 class TestBotRequest:
@@ -19,200 +16,231 @@ class TestBotRequest:
         self.session = tgcli.request.bot.BotSession("0")
         self.request = tgcli.request.bot.BotRequest(self.session, "getMe")
 
-    def test_url(self):
+    def test_url(self, bot_session, bot_request_factory):
+        bot_request = bot_request_factory("getMe")
         url = "{base_url}{method_name}".format(
-            base_url=self.session.base_url, method_name="getMe"
+            base_url=bot_session.base_url, method_name="getMe"
         )
-        assert self.request.url == url
+        assert bot_request.url == url
 
 
 class TestAuthenticationRequest:
-    def setup_method(self):
-        self.adapter = requests_mock.Adapter()
+    def test_url(self, bot_authentication_request):
+        assert bot_authentication_request.url[-5:] == "getMe"
 
-        self.session = tgcli.request.bot.BotSession("0")
-        self.session._is_mocked = True
-
-        self.request = tgcli.request.bot.AuthenticationRequest(self.session)
-
-        self.session.mount("mock", self.adapter)
-
-    def test_url(self):
-        assert self.request.url[-5:] == "getMe"
-
-    def test_found_status(self):
+    def test_found_status(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.1.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=200
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=200,
             )
 
-        response = self.session.send(self.request)
+        response = bot_session.send(bot_authentication_request)
         assert response.status_code == 200
 
-    def test_found_ok(self):
+    def test_found_ok(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.1.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=200
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=200,
             )
 
-        response = self.session.send(self.request)
-        assert response.json().get("ok") == True
+        response = bot_session.send(bot_authentication_request)
+        assert response.json().get("ok")
 
-    def test_found_result(self):
+    def test_found_result(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.1.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=200
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=200,
             )
 
-        response = self.session.send(self.request)
+        response = bot_session.send(bot_authentication_request)
         assert response.json().get("result").get("id") == 1
-        assert response.json().get("result").get("is_bot") == True
+        assert response.json().get("result").get("is_bot")
         assert response.json().get("result").get("first_name") == "FooBot"
         assert response.json().get("result").get("username") == "foobot"
 
-    def test_notfound_status(self):
+    def test_notfound_status(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.2.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=404
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=404,
             )
 
-        response = self.session.send(self.request)
+        response = bot_session.send(bot_authentication_request)
         assert response.status_code == 404
 
-    def test_notfound_ok(self):
+    def test_notfound_ok(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.2.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=404
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=404,
             )
 
-        response = self.session.send(self.request)
-        assert response.json().get("ok") == False
+        response = bot_session.send(bot_authentication_request)
+        assert not response.json().get("ok")
 
-    def test_notfound_error_code(self):
+    def test_notfound_error_code(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.2.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=404
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=404,
             )
 
-        response = self.session.send(self.request)
+        response = bot_session.send(bot_authentication_request)
         assert response.json().get("error_code") == 404
 
-    def test_notfound_description(self):
+    def test_notfound_description(
+        self, mock_adapter, bot_session, bot_authentication_request
+    ):
         with open(
             "tests/resources/responses/bot.TestAuthenticationRequest.2.json"
         ) as f:
-            self.adapter.register_uri(
-                "POST", self.request.url, text=f.read(), status_code=404
+            mock_adapter.register_uri(
+                "POST",
+                bot_authentication_request.url,
+                text=f.read(),
+                status_code=404,
             )
 
-        response = self.session.send(self.request)
+        response = bot_session.send(bot_authentication_request)
         assert response.json().get("description") == "Not Found"
 
 
 class TestSendMessageRequest:
-    def setup_method(self):
-        self.adapter = requests_mock.Adapter()
+    def test_url(self, bot_send_message_request):
+        assert bot_send_message_request.url[-11:] == "sendMessage"
 
-        self.session = tgcli.request.bot.BotSession("0")
-        self.session._is_mocked = True
+    def test_request_body_chat_id(
+        self, bot_send_message_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_message_request)
+        assert request_body.get("chat_id") == 1
 
-        self.request = tgcli.request.bot.SendMessageRequest(
-            self.session, 1, "foo"
-        )
+    def test_request_body_text(
+        self, bot_send_message_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_message_request)
+        assert request_body.get("text") == "foo"
 
-        import json
-
-        self.request_body = json.loads(self.request.body.decode("utf-8"))
-        self.session.mount("mock", self.adapter)
-
-    def test_url(self):
-        assert self.request.url[-11:] == "sendMessage"
-
-    def test_request_body_chat_id(self):
-        assert self.request_body.get("chat_id") == 1
-
-    def test_request_body_text(self):
-        assert self.request_body.get("text") == "foo"
-
-    def test_request_body_parse_mode(self):
-        assert self.request_body.get("parse_mode") == "Markdown"
+    def test_request_body_parse_mode(
+        self, bot_send_message_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_message_request)
+        assert request_body.get("parse_mode") == "Markdown"
 
 
 class TestSendDocumentRequest:
-    def setup_method(self):
-        self.adapter = requests_mock.Adapter()
+    def test_default_url(self, bot_send_document_request):
+        assert bot_send_document_request.url[-12:] == "sendDocument"
 
-        self.session = tgcli.request.bot.BotSession("0")
-        self.session._is_mocked = True
-
-        self.file = open("tests/resources/file.png", "rb")
-        self.request = tgcli.request.bot.SendFileRequest(
-            self.session, 1, self.file, "lorem ipsum"
-        )
-        self.session.mount("mock", self.adapter)
-
-    def teardown_method(self):
-        self.file.close()
-
-    def test_default_url(self):
-        assert self.request.url[-12:] == "sendDocument"
-
-    def test_photo_url(self):
-        session = tgcli.request.bot.BotSession("0")
-        request = tgcli.request.bot.SendFileRequest(
-            self.session,
+    def test_photo_url(self, bot_send_document_request_factory, file_factory):
+        request = bot_send_document_request_factory(
             1,
-            self.file,
+            file_factory("tests/resources/file.png"),
             "lorem ipsum",
             tgcli.request.bot.MediaType.PHOTO,
         )
         assert request.url[-9:] == "sendPhoto"
 
-    def test_audio_url(self):
-        session = tgcli.request.bot.BotSession("0")
-        request = tgcli.request.bot.SendFileRequest(
-            self.session,
+    def test_audio_url(self, bot_send_document_request_factory, file_factory):
+        request = bot_send_document_request_factory(
             1,
-            self.file,
+            file_factory("tests/resources/file.png"),
             "lorem ipsum",
             tgcli.request.bot.MediaType.AUDIO,
         )
         assert request.url[-9:] == "sendAudio"
 
-    def test_video_url(self):
-        session = tgcli.request.bot.BotSession("0")
-        request = tgcli.request.bot.SendFileRequest(
-            self.session,
+    def test_video_url(self, bot_send_document_request_factory, file_factory):
+        request = bot_send_document_request_factory(
             1,
-            self.file,
+            file_factory("tests/resources/file.png"),
             "lorem ipsum",
             tgcli.request.bot.MediaType.VIDEO,
         )
         assert request.url[-9:] == "sendVideo"
 
-    def test_request_body_chat_id(self):
-        assert b"chat_id" in self.request.body
+    def test_request_body_chat_id(self, bot_send_document_request):
+        assert b"chat_id" in bot_send_document_request.body
 
-    def test_request_body_caption(self):
-        assert b"caption" in self.request.body
+    def test_request_body_caption(self, bot_send_document_request):
+        assert b"caption" in bot_send_document_request.body
 
-    def test_request_body_parse_mode(self):
-        assert b"parse_mode" in self.request.body
+    def test_request_body_parse_mode(self, bot_send_document_request):
+        assert b"parse_mode" in bot_send_document_request.body
 
-    def test_request_body_disable_notification(self):
-        assert b"disable_notification" in self.request.body
+    def test_request_body_disable_notification(
+        self, bot_send_document_request
+    ):
+        assert b"disable_notification" in bot_send_document_request.body
 
-    def test_request_body_document(self):
-        assert b'filename="file.png"' in self.request.body
+    def test_request_body_document(self, bot_send_document_request):
+        assert b'filename="file.png"' in bot_send_document_request.body
+
+
+class TestSendPollRequest:
+    def test_url(self, bot_send_poll_request):
+        assert bot_send_poll_request.url[-8:] == "sendPoll"
+
+    def test_request_body_chat_id(
+        self, bot_send_poll_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_poll_request)
+        assert request_body.get("chat_id") == 1
+
+    def test_request_body_question(
+        self, bot_send_poll_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_poll_request)
+        assert request_body.get("question") == "Foo?"
+
+    def test_request_body_options(
+        self, bot_send_poll_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_poll_request)
+        assert request_body.get("options")[0] == "Bar"
+        assert request_body.get("options")[1] == "Baz"
+
+    def test_request_body_disable_notification(
+        self, bot_send_poll_request, request_body_factory
+    ):
+        request_body = request_body_factory(bot_send_poll_request)
+        assert not request_body.get("disable_notification")