+++ tgcli/__init__.py
@@ -1,2 +1,2 @@
 __author__ = "Eray Erdin"
-__version__ = "0.1.2"
+__version__ = "0.2.0"